[Silver III] Title: 주유소, Time: 400 ms, Memory: 36388 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,24 +7,24 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
         
-        int[] distance = new int[N-1];
-        int[] cost = new int[N];
+        long[] distance = new long[N-1];
+        long[] cost = new long[N];
         
         StringTokenizer st = new StringTokenizer(br.readLine());
         
         for (int i = 0; i < N - 1; i++) {
-			distance[i] = Integer.parseInt(st.nextToken());
+			distance[i] = Long.parseLong(st.nextToken());
 		}
         
         st = new StringTokenizer(br.readLine());
         
         for (int i = 0; i < N ; i++) {
-			cost[i] = Integer.parseInt(st.nextToken());
+			cost[i] = Long.parseLong(st.nextToken());
 		}
         
         
-        int min = cost[0];
-        int ans = 0;
+        long min = cost[0];
+        long ans = 0;
         for (int i = 0; i < distance.length; i++) {
 			if( cost[i] < min) {
 				min = cost[i];
